+++ src/main/java/mirsario/cameraoverhaul/utilities/VectorUtils.java
@@ -1,47 +1,47 @@
 package mirsario.cameraoverhaul.utilities;
 
-import org.joml.Vector2f;
-import org.joml.Vector3f;
+import org.joml.Vector2d;
+import org.joml.Vector3d;
 
 public final class VectorUtils {
 	#if MC_RELEASE
-	public static Vector2f toJoml(net.minecraft.world.phys.Vec2 vec) {
-		return new Vector2f((float)vec.x, (float)vec.y);
+	public static Vector2d toJoml(net.minecraft.world.phys.Vec2 vec) {
+		return new Vector2d(vec.x, vec.y);
 	}
-	public static Vector3f toJoml(net.minecraft.world.phys.Vec3 vec) {
-		return new Vector3f((float)vec.x, (float)vec.y, (float)vec.z);
+	public static Vector3d toJoml(net.minecraft.world.phys.Vec3 vec) {
+		return new Vector3d(vec.x, vec.y, vec.z);
 	}
 	#endif
 
-	public static float length(Vector2f vec) {
+	public static double length(Vector2d vec) {
 		return length(vec.x, vec.y);
 	}
-	public static float length(float x, float y) {
-		return (float)Math.sqrt(x * x + y * y);
+	public static double length(double x, double y) {
+		return Math.sqrt(x * x + y * y);
 	}
 
-	public static Vector2f rotate(Vector2f vec, float degrees) {
+	public static Vector2d rotate(Vector2d vec, double degrees) {
 		double radians = Math.toRadians(degrees);
-		float sin = (float)Math.sin(radians);
-		float cos = (float)Math.cos(radians);
+		double sin = Math.sin(radians);
+		double cos = Math.cos(radians);
 
-		return new Vector2f((cos * vec.x) - (sin * vec.y), (sin * vec.x) + (cos * vec.y));
+		return new Vector2d((cos * vec.x) - (sin * vec.y), (sin * vec.x) + (cos * vec.y));
 	}
 
-	public static Vector2f lerp(Vector2f src, Vector2f dst, float step) {
+	public static Vector2d lerp(Vector2d src, Vector2d dst, double step) {
 		return lerp(src.x, src.y, dst.x, dst.y, step);
 	}
-	public static Vector2f lerp(float xSrc, float ySrc, float xDst, float yDst, float step) {
-		return new Vector2f(
+	public static Vector2d lerp(double xSrc, double ySrc, double xDst, double yDst, double step) {
+		return new Vector2d(
 			xSrc + (xDst - xSrc) * step,
 			ySrc + (yDst - ySrc) * step
 		);
 	}
 
-	public static Vector2f multiply(Vector2f vec, float value) {
-		return new Vector2f(vec.x * value, vec.y * value);
+	public static Vector2d multiply(Vector2d vec, double value) {
+		return new Vector2d(vec.x * value, vec.y * value);
 	}
-	public static Vector2f multiply(Vector2f vec, Vector2f value) {
-		return new Vector2f(vec.x * value.x, vec.y * value.y);
+	public static Vector2d multiply(Vector2d vec, Vector2d value) {
+		return new Vector2d(vec.x * value.x, vec.y * value.y);
 	}
 }
